- Add functionality for guest user - Increase delete button size - Move some text from layout to string file - Add confirm dialog to delete task - Optimize language option

+++ app/src/main/java/com/example/tasktimerappkk/TasksActivity.kt
@@ -1,13 +1,15 @@
 package com.example.tasktimerappkk
 
-import android.content.BroadcastReceiver
-import android.content.Context
-import android.content.Intent
-import android.content.IntentFilter
+import android.content.*
+import android.graphics.Color
 import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
+import android.widget.TextView
+import android.widget.Toast
+import androidx.appcompat.app.AlertDialog
 import androidx.lifecycle.ViewModelProvider
 import com.example.tasktimerappkk.Model.Task
+import com.example.tasktimerappkk.Model.TaskL
 import com.example.tasktimerappkk.ViewModel.MyViewModel
 import com.example.tasktimerappkk.databinding.ActivityTasksBinding
 import kotlin.math.roundToInt
@@ -31,17 +33,30 @@ import kotlin.math.roundToInt
             context=this
 
             binding.apply {
+                //If the user login then we will get data from firebase, else the data will retrive from room
 
-
-                viewModel = ViewModelProvider(this@TasksActivity).get(MyViewModel::class.java)
-                //Read data from database
-                viewModel.getTasks().observe(this@TasksActivity, { userTasks ->
-                    TasksAdapter.UpdateList(userTasks)
-                    updateTotal(userTasks)
-                })
-
+                //**************************************************************
+               // viewModel = ViewModelProvider(this@TasksActivity).get(MyViewModel::class.java)
                 TasksAdapter = TasksAdapter(this@TasksActivity)
                 binding.tasksRv.adapter = TasksAdapter
+                if (MainActivity.userData.user!=null) {
+                    //Read data from database
+                    viewModel.getTasks().observe(this@TasksActivity) { userTasks ->
+                        TasksAdapter.UpdateList(userTasks)
+                        updateTotal(userTasks)
+                    }
+
+
+                }
+                else{
+
+                    viewModel.getLocalTasks().observe(this@TasksActivity) { localTasks ->
+                        TasksAdapter.UpdateLocalList(localTasks)
+                        updateLocalTotal(localTasks)
+
+                    }
+                }
+                //****************************************************************
                 addingBtn.setOnClickListener {
                     var intent = Intent(context, AddTaskActivity::class.java)
                     context.startActivity(intent)
@@ -51,8 +66,23 @@ import kotlin.math.roundToInt
         }
 
 
-        override fun deleteTimer(task: Task) {
-            viewModel.deleteTask(task)
+        override fun deleteTask(task: Task) {
+
+            val alertDialog = AlertDialog.Builder(this)
+            val title = TextView(this)
+            title.setTextColor(Color.BLACK)
+            title.setText(getString(R.string.deleteConf))
+            alertDialog.setView(title)
+
+            alertDialog.setPositiveButton(getString(R.string.delete)) { dialog, i ->
+                viewModel.deleteTask(task)
+                Toast.makeText(this, getString(R.string.deleted), Toast.LENGTH_LONG).show()
+
+            }//end delete option
+                .setNegativeButton(getString(R.string.cancel), DialogInterface.OnClickListener {
+                        dialog, id -> dialog.cancel()
+                })
+            alertDialog.show()
         }
 
         override fun updateTotal(tasks: List<Task>) {
@@ -62,56 +92,44 @@ import kotlin.math.roundToInt
                 }
             binding.totalTimeTv.text=getTimeStringFromDouble(totalTime)
         }
+        //------------------------------------------------------------
+        override fun updateLocalTotal(tasks: List<TaskL>) {
+            totalTime=0.0
+            for(i in tasks){
+                totalTime+=i.timer
+            }
+            binding.totalTimeTv.text=getTimeStringFromDouble(totalTime)
+        }
+        override fun deleteLocalTask(task: TaskL) {
+
+            val alertDialog = AlertDialog.Builder(this)
+            val title = TextView(this)
+            title.setTextColor(Color.BLACK)
+            title.setText(getString(R.string.deleteConf))
+            alertDialog.setView(title)
+
+            alertDialog.setPositiveButton(getString(R.string.delete)) { dialog, i ->
+                viewModel.deleteLocalTask(task)
+                Toast.makeText(this, getString(R.string.deleted), Toast.LENGTH_LONG).show()
+
+            }//end delete option
+                .setNegativeButton(getString(R.string.cancel), DialogInterface.OnClickListener {
+                        dialog, id -> dialog.cancel()
+                })
+            alertDialog.show()
+        }
+        override fun updateLocalTimer(task: TaskL, time:Double){
+            task.timer=time
+
+            viewModel.editLocalTask(task,time.toString())
+
+        }
+        //------------------------------------------------------------
 
         override fun updateTimer(task: Task, time: Double) {
             viewModel.editTask(task,time.toString())
         }
 
-        /*
-                override fun deleteTimer(task: Task) {
-                    viewModel.deleteTask(task)
-                }
-
-            override fun updateTotal() {
-                binding.totalTimeTv.text=getTimeStringFromDouble(totalTime)
-            }
-
-            private fun resetTimer() {
-                    stopTimer()
-                    time=0.0
-                }
-
-                private fun startStopTimer() {
-                    if(timerStarted)
-                        stopTimer()
-                    else
-                        startTimer()
-                }
-
-                private fun startTimer() {
-                    var binding= TasksRowBinding.inflate(layoutInflater)
-                    binding.timerBtn.setImageResource(R.drawable.ic_baseline_pause_circle_filled_24)
-                    serviceIntent.putExtra(TimerService.TIME_EXTRA,time)
-                    startService(serviceIntent)
-                    timerStarted=true
-                }
-
-                private fun stopTimer() {
-                    var binding= TasksRowBinding.inflate(layoutInflater)
-                    stopService(serviceIntent)
-                    binding.timerBtn.setImageResource(R.drawable.started_icon)
-                    timerStarted=false
-                }
-
-
-                private val  updateTime : BroadcastReceiver = object : BroadcastReceiver() {
-                    override fun onReceive(context: Context?, intent: Intent?) {
-                        var binding= TasksRowBinding.inflate(layoutInflater)
-                        time=intent!!.getDoubleExtra(TimerService.TIME_EXTRA,0.0)
-                        binding.timerTv.text=getTimeStringFromDouble(time)
-                    }
-
-                }*/
 
             private fun getTimeStringFromDouble(time: Double): String {
                 val resultInt = time.roundToInt()
